ENH: Put plugin/packages panel in its own tab
